fix: reorganize admin deal edit route
